# FIX: improve performance of UserStat.ensure_consistency